requirements.txt
# ============================================================
# ETF Bot için Gerekli Kütüphaneler
# ============================================================
# Her satır bir Python paketi, pip install ile yüklenir.
# Kullanım: pip install -r requirements.txt
# ============================================================

pandas>=2.0.0
numpy>=1.24.0
gymnasium>=0.29.0
stable-baselines3>=2.0.0
fredapi>=0.5.0
yfinance>=0.2.0
matplotlib>=3.7.0
python-dotenv>=1.0.0
ib_insync>=0.9.0
torch>=2.0.0
